requirements.txt
aiohttp
psutil
psycopg2-binary
sqlalchemy
asyncio
asyncpg
greenlet
alembic
pyyaml